chore: move ServerAPI to server-api
Fixes #1516
